DB model v1 + Simple Firebase Log in mechanism
I`ve created a first Firebase data model (i might change it later on), and tested in TaskFragment a code that actually inputs data in each child of the root node. This code i will later on use for the admin and client app.

Additionally i`ve created a simple Firebase login which will be used as a base for futher enhacements.

ToDo: Try to change the library versions